requirements.txt
paho-mqtt
requests
holidays
pytz
